Reorganize files and imports.
