+++ src/evaluate.jl
@@ -1,4 +1,11 @@
-export Configuration
+export evaluate, interactive_shell
+
+using Dates
+using Random
+using DataFrames: DataFrame, nrow
+using LazyArtifacts: @artifact_str
+using ProgressMeter: Progress, update!
+using Sandbox: Sandbox, SandboxConfig, UnprivilegedUserNamespacesExecutor, cleanup
 
 lazy_artifact(x) = @artifact_str(x)
